권한 확인 로직 추가하여 product 페이지 접근 제한

+++ src/main/java/rlguswn/JavaShop/controller/ProductController.java
@@ -1,11 +1,13 @@
 package rlguswn.JavaShop.controller;
 
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import rlguswn.JavaShop.domain.Product;
 import rlguswn.JavaShop.dto.product.ProductRegisterForm;
 import rlguswn.JavaShop.dto.product.ProductUpdateForm;
+import rlguswn.JavaShop.service.MemberService;
 import rlguswn.JavaShop.service.ProductImageService;
 import rlguswn.JavaShop.service.ProductService;
 
@@ -15,10 +17,12 @@ import java.util.List;
 @RequestMapping("/product")
 public class ProductController {
 
+    private final MemberService memberService;
     private final ProductService productService;
     private final ProductImageService productImageService;
 
-    public ProductController(ProductService productService, ProductImageService productImageService) {
+    public ProductController(MemberService memberService, ProductService productService, ProductImageService productImageService) {
+        this.memberService = memberService;
         this.productService = productService;
         this.productImageService = productImageService;
     }
@@ -31,12 +35,18 @@ public class ProductController {
     }
 
     @GetMapping("/register")
-    public String registerPage() {
+    public String registerPage(Model model) {
+        if (!memberService.getLoginMember().isAdmin()) {
+            throw new AccessDeniedException("접근 권한이 없습니다.");
+        }
         return "product/productRegister";
     }
 
     @PostMapping(value = "/register", consumes = "multipart/form-data")
     public String registerProduct(@ModelAttribute ProductRegisterForm form) {
+        if (!memberService.getLoginMember().isAdmin()) {
+            throw new AccessDeniedException("접근 권한이 없습니다.");
+        }
         Product product = productService.registerProduct(form);
         Long productId = product.getId();
         return "redirect:/product/" + productId;
@@ -51,6 +61,9 @@ public class ProductController {
 
     @GetMapping("/{id}/update")
     public String updatePage(Model model, @PathVariable Long id) {
+        if (!memberService.getLoginMember().isAdmin()) {
+            throw new AccessDeniedException("접근 권한이 없습니다.");
+        }
         Product product = productService.getProductById(id).get();
         model.addAttribute("product", product);
         return "product/productUpdate";
@@ -58,6 +71,9 @@ public class ProductController {
 
     @PostMapping(value = "/{id}/update", consumes = "multipart/form-data")
     public String updateProduct(Model model, @PathVariable Long id, @ModelAttribute ProductUpdateForm form) {
+        if (!memberService.getLoginMember().isAdmin()) {
+            throw new AccessDeniedException("접근 권한이 없습니다.");
+        }
         Product product = productService.updateProduct(id, form);
         Long productId = product.getId();
         model.addAttribute("product", product);
@@ -66,6 +82,9 @@ public class ProductController {
 
     @GetMapping("/{id}/delete")
     public String deleteProduct(@PathVariable Long id) {
+        if (!memberService.getLoginMember().isAdmin()) {
+            throw new AccessDeniedException("접근 권한이 없습니다.");
+        }
         productService.deleteById(id);
         return "redirect:/product";
     }
